gcs-to-supabase-sync/requirements.txt
functions-framework==3.*
cloudevents==1.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
supabase==2.*

